Changed the behaviour of site images.
A difficult design decision was made today. Images, instead of loading
with the GET of a page, are now loaded after the user clicks on a
thumbnail, and they are opened in a seperate tab. This is not the
optimal way of browsing an image board, but unfortunately, this is the
only way that I am currently aware of to prevent the page from loading
20MB+ of images by default, without using jQuery or some JavaScript
fuckery that is antithetical to my values.
userscripts can add more fluid functionality, but fort the moment, this
is the way that the default package will be distributed.
Also the UNSCII font from omnius was removed.